+++ packages/article-extras/src/article-extras.js
@@ -36,7 +36,8 @@ const ArticleExtras = ({
   isCommentEnabled,
   storefrontConfig,
   breadcrumbs,
-  domainSpecificUrl
+  domainSpecificUrl,
+  isEntitlementFeatureEnabled
 }) => {
   const renderBreadcrumb = ({ showBorder } = { showBorder: false }) => {
     if (breadcrumbs && breadcrumbs.length > 0) {
@@ -114,6 +115,7 @@ const ArticleExtras = ({
         isCommentEnabled={isCommentEnabled}
         storefrontConfig={storefrontConfig}
         domainSpecificUrl={domainSpecificUrl}
+        isEntitlementFeatureEnabled={isEntitlementFeatureEnabled}
       />
     </UserState>
   );
@@ -139,7 +141,8 @@ ArticleExtras.propTypes = {
   isCommentEnabled: PropTypes.bool,
   storefrontConfig: PropTypes.string.isRequired,
   breadcrumbs: PropTypes.arrayOf(PropTypes.shape({})),
-  domainSpecificUrl: PropTypes.string.isRequired
+  domainSpecificUrl: PropTypes.string.isRequired,
+  isEntitlementFeatureEnabled: PropTypes.bool.isRequired
 };
 
 ArticleExtras.defaultProps = {