+++ packages/article-in-depth/fixtures/full-article.js
@@ -1850,6 +1850,8 @@ const defaultRelatedArticleSlice = {
             name: "paragraph"
           }
         ],
+        categoryPath:
+          "/article/rise-of-centenarian-drivers-as-elderly-push-on-6gb0bjnpz",
         url:
           "http://cps-render-ci.elb.tnl-dev.ntch.co.uk/article/rise-of-centenarian-drivers-as-elderly-push-on-6gb0bjnpz"
       }
@@ -1995,6 +1997,8 @@ const defaultRelatedArticleSlice = {
             name: "paragraph"
           }
         ],
+        categoryPath:
+          "/article/youtube-ignored-police-pleas-to-remove-threatening-videos-62v2mqp23",
         url:
           "https://www.thetimes.co.uk/article/youtube-ignored-police-pleas-to-remove-threatening-videos-62v2mqp23"
       }