+++ packages/article-skeleton/fixtures/full-article.js
@@ -1312,6 +1312,7 @@ const defaultRelatedArticleSlice = {
             name: "paragraph"
           }
         ],
+        categoryPath: "/article/tms-pratchetts-law-of-the-jungle-xgqrcw779",
         url:
           "http://cps-render-ci.elb.tnl-dev.ntch.co.uk/article/tms-pratchetts-law-of-the-jungle-xgqrcw779"
       }
@@ -1504,6 +1505,8 @@ const defaultRelatedArticleSlice = {
             name: "paragraph"
           }
         ],
+        categoryPath:
+          "/article/rise-of-centenarian-drivers-as-elderly-push-on-6gb0bjnpz",
         url:
           "http://cps-render-ci.elb.tnl-dev.ntch.co.uk/article/rise-of-centenarian-drivers-as-elderly-push-on-6gb0bjnpz"
       }
@@ -1649,6 +1652,8 @@ const defaultRelatedArticleSlice = {
             name: "paragraph"
           }
         ],
+        categoryPath:
+          "/article/youtube-ignored-police-pleas-to-remove-threatening-videos-62v2mqp23",
         url:
           "https://www.thetimes.co.uk/article/youtube-ignored-police-pleas-to-remove-threatening-videos-62v2mqp23"
       }