+++ packages/article-skeleton/src/article-skeleton.js
@@ -87,7 +87,8 @@ const ArticleSkeleton = ({
     publishedTime,
     isSavingEnabled,
     isSharingEnabled,
-    isCommentEnabled
+    isCommentEnabled,
+    isEntitlementFeatureEnabled
   } = article;
 
   const [showVerifyEmailBanner, setShowEmailVerifyBanner] = useState(false);
@@ -430,6 +431,7 @@ const ArticleSkeleton = ({
                       storefrontConfig={storefrontConfig}
                       breadcrumbs={breadcrumbs}
                       domainSpecificUrl={domainSpecificUrl}
+                      isEntitlementFeatureEnabled={isEntitlementFeatureEnabled}
                     />
                   )}
                 </LazyLoad>