+++ packages/ts-components/src/components/save-star/SaveStar.tsx
@@ -1,4 +1,10 @@
-import React, { cloneElement, useCallback, useMemo, useState } from 'react';
+import React, {
+  cloneElement,
+  useCallback,
+  useEffect,
+  useMemo,
+  useState
+} from 'react';
 import { FetchProvider } from '../../helpers/fetch/FetchProvider';
 import { SaveStarUI } from './SaveStarUI';
 import { ContentProvider } from './ContentProvider';
@@ -8,8 +14,15 @@ export const SaveStar: React.FC<{
   articleId: string;
   isPreviewMode?: boolean;
 }> = React.memo(({ articleId, isPreviewMode, children }) => {
-  const [url, setUrl] = useState<string>(
-    `/api/collections/is-bookmarked/${articleId}`
+  const [url, setUrl] = useState<string | undefined>();
+
+  useEffect(
+    () => {
+      if (window.nuk && window.nuk.user && window.nuk.user.isLoggedIn) {
+        setUrl(`/api/collections/is-bookmarked/${articleId}`);
+      }
+    },
+    [articleId]
   );
 
   const [previewData, setPreviewData] = useState<ArticleBookmark | undefined>(