+++ packages/ts-components/src/index.ts
@@ -91,6 +91,7 @@ export {
 
 // Helpers
 export { FetchProvider } from './helpers/fetch/FetchProvider';
+
 export {
   ViewCountWrapper
 } from './helpers/view-count-wrapper/ViewCountWrapper';